refactor: paswword -> password 오타 lender -> render 오타 수정
