+++ src/components/editprofile/Edit.tsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import EditPaswword from "./EditPaswword";
+import EditPassword from "./EditPassword";
 import EditProfile from "./EditProfile";
 import { useEffect, useState } from "react";
 import { auth } from "@/firebase/app";
@@ -29,7 +29,7 @@ export default function Edit() {
       <EditButton onClick={PasswordClcikHandler} isClick={click ? false : true}>
         비밀번호 변경
       </EditButton>
-      {!click ? <EditProfile /> : <EditPaswword />}
+      {!click ? <EditProfile /> : <EditPassword />}
     </Section>
   );
 }
